refactor: 控制台重构成两个端，并把这条边界真的落到服务端
前端的菜单和路由原来是运行时拼出来的：后端返回一棵 permission 树，每行带
path、component（上一版 Vue 前端的文件路径）、icon、order，前端据此建路由和
tab，菜单项再按 code 过滤。于是"页面存在"和"菜单可见"由两个地方分别决定，可以
互相打架——新增一个页面必须同时往 permission 表补一条记录，漏了就是对谁都不可
见，包括超级管理员（它拿到的是"所有顶级记录"，不是"全部权限"）。

现在只有两个端，路由写死在前端代码里：

| 端     | 路径       | 谁进得来      |
|--------|-----------|--------------|
| 管理端 | /admin/*  | SUPER_ADMIN  |
| 用户端 | /portal/* | 其余所有角色  |

分流点只有一条：identity.Principal.IsSuperAdmin()。后端本来也只有这一条授权边
界，所以端不多不少正好两个；角色除了决定进哪个端，不再有别的作用。

用户端在 /portal 而不是 /user：/user 是账号接口的前缀，Ingress、nginx.conf 和
vite dev proxy 都会把它整个路由到后端，/user/profile 会打到 API 而不是前端。
e2e 脚本里加了断言盯着这一点。

删掉的东西：permission / role_permissions_permission 两张表、/permission/* 整
组接口、角色的增删改与分配用户、前端的 lib/permissions.ts 与 routes/nav.tsx。
已有的库用 sql/migrate-2026-09-05-two-sides.sql 清理（在真实 MySQL 8 上验证过：
可重入，只挂 ROLE_QA 的账号会先归还给 USER）。

顺带把这条边界真的做成边界。两个端由同一套 API 提供服务，用户端账号手里的
token 对每条路由都合法，所以"用户端界面上没有这个入口"从来就不是限制：

- /auth/current-role/switch/:role 不校验目标角色，任何 token 都能一次请求换来
  SUPER_ADMIN——中间件会把 currentRoleCode 原样搬进 Principal，上面整套分端和
  service 层的权限判断就全空了
- 账号那组路由（列表/新增/改/删/重置密码/GET /role）只有 Jwt，没有任何授权检查，
  谁都能给自己 PATCH 一个 SUPER_ADMIN 角色
- ListUsers / EditUser / AssociatePlan / ListPlans / 链路组的写接口都没有检查，
  其中 AssociatePlan 能让任何账号给自己挂上最大的套餐外加一个新的到期时间
- ModifyRule / ModifyRules / TestRule 不校验规则归属，改 remote_addr 就能把别人
  的转发悄悄指到自己的机器上（DelRule 一直是查的，这三个是漏的）
- entity.Node / entity.Chain 会把 key 一起序列化，那是控制面命令 agent 用的凭证；
  前端不显示这两列不等于它们没发出去

对应加了 middleware.SuperAdmin（放在路由表里，读 admin_route.go 就知道哪些接口
属于哪个端）、service.requireSuperAdmin、checkRuleOwnership 和 redactForCaller。
个人资料按 profile 行的归属放行，两个端都能改自己的。

另外修了两个真实存在的显示错误：

- GetUserPlanInfo 的用户 id 直接取自请求体，任何账号改个数字就能读到别人的套餐
  和流量历史；现在非管理员一律按自己算
- 流量单位不一致：GetUserPlanInfo 和 ListUsers 返回的是整数 GB，旁边的
  traffic_plan.total_traffic 是字节，前端两边都用 formatBytes——5 GB 显示成
  "5 B"，不到 1 GB 的一律显示 0，转发用户页的使用率进度条恒为 0%。统一成字节。

主题默认改成亮色，深色和右上角的切换保留。默认是明确的 light 而不是 system：
系统设成深色的人不该在没选过的情况下拿到一个深色控制台。

+++ internal/service/resp/response.go
@@ -89,19 +89,26 @@ type ListNodeResp struct {
 type ReportNetworkTrafficResp struct {
 }
 
+// GetUserPlanInfoResp reports one user's plan and consumption.
+//
+// Every traffic figure here is a byte count, like every other traffic figure
+// the API returns. These three used to be pre-divided into whole gigabytes,
+// which threw away everything below 1 GB — a user who had spent 800 MB was
+// shown 0 — and left the frontend formatting a gigabyte count with a
+// byte formatter, so 5 GB rendered as "5 B".
 type GetUserPlanInfoResp struct {
 	Username         string
 	ExpirationTime   time.Time
 	ResetTrafficTime time.Time
-	PlanTraffic      int32
-	UsedTraffic      int32
+	PlanTraffic      int64
+	UsedTraffic      int64
 	PlanName         string
 	DailyTraffics    []DailyTraffic
 }
 
 type DailyTraffic struct {
 	Date    time.Time
-	Traffic int32
+	Traffic int64
 }
 
 type EditChainResp struct {
